+++ mobile/android/android-components/components/browser/awesomebar/src/main/java/mozilla/components/browser/awesomebar/layout/DefaultSuggestionViewHolder.kt
@@ -2,44 +2,35 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.browser.awesomebar
+package mozilla.components.browser.awesomebar.layout
 
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import mozilla.components.browser.awesomebar.layout.FlowLayout
+import mozilla.components.browser.awesomebar.BrowserAwesomeBar
+import mozilla.components.browser.awesomebar.R
+import mozilla.components.browser.awesomebar.widget.FlowLayout
 import mozilla.components.concept.awesomebar.AwesomeBar
 
-/**
- * [RecyclerView.ViewHolder] implementations.
- */
-internal sealed class SuggestionViewHolder(
-    itemView: View
-) : RecyclerView.ViewHolder(itemView) {
-    /**
-     * Binds the views in the holder to the passed suggestion.
-     */
-    abstract fun bind(suggestion: AwesomeBar.Suggestion)
-
+internal sealed class DefaultSuggestionViewHolder {
     /**
      * Default view holder for suggestions.
      */
-    internal class DefaultSuggestionViewHolder(
+    internal class Default(
         private val awesomeBar: BrowserAwesomeBar,
-        itemView: View
-    ) : SuggestionViewHolder(itemView) {
-        private val titleView = itemView.findViewById<TextView>(R.id.mozac_browser_awesomebar_title).apply {
+        view: View
+    ) : SuggestionViewHolder(view) {
+        private val titleView = view.findViewById<TextView>(R.id.mozac_browser_awesomebar_title).apply {
             setTextColor(awesomeBar.styling.titleTextColor)
         }
-        private val descriptionView = itemView.findViewById<TextView>(R.id.mozac_browser_awesomebar_description).apply {
+        private val descriptionView = view.findViewById<TextView>(R.id.mozac_browser_awesomebar_description).apply {
             setTextColor(awesomeBar.styling.descriptionTextColor)
         }
-        private val iconView = itemView.findViewById<ImageView>(R.id.mozac_browser_awesomebar_icon)
+        private val iconView = view.findViewById<ImageView>(R.id.mozac_browser_awesomebar_icon)
 
-        override fun bind(suggestion: AwesomeBar.Suggestion) {
+        override fun bind(suggestion: AwesomeBar.Suggestion, selectionListener: () -> Unit) {
             val title = if (suggestion.title.isNullOrEmpty()) suggestion.description else suggestion.title
 
             val icon = suggestion.icon.invoke(iconView.measuredWidth, iconView.measuredHeight)
@@ -48,9 +39,9 @@ internal sealed class SuggestionViewHolder(
             titleView.text = title
             descriptionView.text = suggestion.description
 
-            itemView.setOnClickListener {
+            view.setOnClickListener {
                 suggestion.onSuggestionClicked?.invoke()
-                awesomeBar.listener?.invoke()
+                selectionListener.invoke()
             }
         }
 
@@ -62,19 +53,19 @@ internal sealed class SuggestionViewHolder(
     /**
      * View holder for suggestions that contain chips.
      */
-    internal class ChipsSuggestionViewHolder(
+    internal class Chips(
         private val awesomeBar: BrowserAwesomeBar,
-        itemView: View
-    ) : SuggestionViewHolder(itemView) {
-        private val iconView = itemView.findViewById<ImageView>(R.id.mozac_browser_awesomebar_icon)
-        private val chipsView = itemView.findViewById<FlowLayout>(R.id.mozac_browser_awesomebar_chips).apply {
+        view: View
+    ) : SuggestionViewHolder(view) {
+        private val iconView = view.findViewById<ImageView>(R.id.mozac_browser_awesomebar_icon)
+        private val chipsView = view.findViewById<FlowLayout>(R.id.mozac_browser_awesomebar_chips).apply {
             spacing = awesomeBar.styling.chipSpacing
         }
 
-        override fun bind(suggestion: AwesomeBar.Suggestion) {
+        override fun bind(suggestion: AwesomeBar.Suggestion, selectionListener: () -> Unit) {
             chipsView.removeAllViews()
 
-            val inflater = LayoutInflater.from(itemView.context)
+            val inflater = LayoutInflater.from(view.context)
 
             suggestion.icon.invoke(iconView.measuredWidth, iconView.measuredHeight)?.let { bitmap ->
                 iconView.setImageBitmap(bitmap)
@@ -85,7 +76,7 @@ internal sealed class SuggestionViewHolder(
                 .forEach { chip ->
                     val view = inflater.inflate(
                         R.layout.mozac_browser_awesomebar_chip,
-                        itemView as ViewGroup,
+                        view as ViewGroup,
                         false
                     ) as TextView
 
@@ -94,7 +85,7 @@ internal sealed class SuggestionViewHolder(
                     view.text = chip.title
                     view.setOnClickListener {
                         suggestion.onChipClicked?.invoke(chip)
-                        awesomeBar.listener?.invoke()
+                        selectionListener.invoke()
                     }
 
                     chipsView.addView(view)