+++ mobile/android/android-components/components/browser/awesomebar/src/test/java/mozilla/components/browser/awesomebar/widget/FlowLayoutTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.browser.awesomebar.layout
+package mozilla.components.browser.awesomebar.widget
 
 import android.content.Context
 import android.view.View